In the old version, the names of the files to read were hard coded, relative to the root of the Mu2e Offline package. And the code called ConfigFileLookupPolicy to turn the hard coded name into a complete path, relative to MU2E_SEARCH_PATH.
In the new version, this code talks to Mu2e supplied code via
BaBar/inc/ExternalInfo.hh, a singleton-like class that provides access to
Mu2e supplied information.  The full path names are obtained from this
service.  The names of the files are now supplied in the configuration
of the Mu2e BtrkHelper service.  See item 8 in the release notes for Mu2e
Offline v5_3_4.